Code cleanup to fix warnings on Windows and macOS builds, added -dump-memory option (#304)
* Add defaults to switch statements

* CReal3D: Fixed warnings, moved default initialization from constructor to class declaration

* CLegacy3D: Fix warnings

* CNew3D: fixed warnings about member order initialization in constructor

* Fixed warnings

* New3D engine: fixed offsetof() warnings

* PowerPC disassembler: sprintf() -> snprintf() and bumped its version number up

* Added -dump-memory option

* ppcd: Fixed potential segfault

* Input.cpp: fix constructor initializer list order

* 53C810.cpp: Fixed warning about extraneous parentheses

* Update copyright headers

* Z80.cpp: Remove pointless self assignments (were these ANDed with 0xffff at some point?)

* Updated copyright header

* Musashi: Fixed warnings

* Update copyright header

* SCSP.cpp: Removed non-functional placeholders

* Main.cpp: Commented out debug callback

* ppc.cpp: ppc_set_cr() for entire CR register was never used

* DSB.cpp: stateName table only used by a commented-out printf, turned that into a DebugLog

* New3D engine: Fixed warnings in R3DScrollFog

* GameLoader.cpp: Fix warnings

* Audio.cpp: Fixed warnings

* Pragma is MSVC-specific

* Makefile.OSX: Suppress deprecation warnings for gluPerspective()

* Logger.cpp: fixed warning about format string not being a string literal when using syslog()

* Makefile.OSX: remove -s option from link step

* CModel3: Fixed sign mismatch warning

* ppc_ops.c: Fix FP environment pragmas to support MSVC, clang, and gcc properly

* ppc.cpp: Fixed sign warning

* DirectInputSystem: Fixed various warnings

* SDLInputSystem: Fix sign warnings

* Input.cpp: Fix string assembly warning

* Fixed CReal3D warnings about sentinel sign and fixed a bit cast

* Main.cpp: Fixed sign warnings

* Makefiles/Rules.inc: Disable warnings for glew.c, which is a third-party file

* SCSP.cpp: Fixed type punning error

* Ian's preferred fix for offsetof(): composition with Vertex

+++ Src/OSD/SDL/SDLInputSystem.cpp
@@ -468,7 +468,7 @@ bool CSDLInputSystem::InitializeSystem()
 
 int CSDLInputSystem::GetKeyIndex(const char *keyName)
 {
-  for (int i = 0; i < NUM_SDL_KEYS; i++)
+  for (size_t i = 0; i < NUM_SDL_KEYS; i++)
   {
     if (stricmp(keyName, s_keyMap[i].keyName) == 0)
       return i;
@@ -478,7 +478,7 @@ int CSDLInputSystem::GetKeyIndex(const char *keyName)
 
 const char *CSDLInputSystem::GetKeyName(int keyIndex)
 {
-  if (keyIndex < 0 || keyIndex >= NUM_SDL_KEYS)
+  if (keyIndex < 0 || keyIndex >= (int)NUM_SDL_KEYS)
     return nullptr;
   return s_keyMap[keyIndex].keyName;
 }
